feat: added API communication between pipelines and airflow services (#737)

+++ pipelines/pipelines/app.py
@@ -1,10 +1,17 @@
 import asyncio
+import datetime
 import os
+from math import ceil
 from typing import Any, Dict, List, Optional, Union
 
+from airflow_client.client.api.dag_api import DAGApi
+from airflow_client.client import ApiException, ApiClient as AirflowClient
+from airflow_client.client.api.dag_run_api import DAGRunApi
+from airflow_client.client.model.dag_run import DAGRun
 from fastapi import Depends, FastAPI, Header, HTTPException, status
 from fastapi.middleware.cors import CORSMiddleware
 from filter_lib import Page, form_query, map_request_to_filter, paginate
+from filter_lib.pagination import PaginationParams
 from pydantic import AnyUrl
 from sqlalchemy.orm import Session
 from sqlalchemy_filters.exceptions import BadFilterFormat
@@ -13,6 +20,7 @@ from tenant_dependency import TenantData, get_tenant_info
 import pipelines.config as config
 import pipelines.db.models as dbm
 import pipelines.db.service as service
+import pipelines.airflow_utils as airflow_utils
 import pipelines.execution as execution
 import pipelines.schemas as schemas
 from pipelines.kafka_utils import Kafka
@@ -69,7 +77,14 @@ async def add_pipeline(
     session: Session = Depends(service.get_session),
 ) -> schemas.PipelineOutId:
     """Add pipeline to DB."""
-    # TODO: fix
+    # TODO: fix model cluster
+
+    if pipeline.meta.name.endswith(':airflow'):
+        raise HTTPException(
+            status_code=400,
+            detail='Airflow pipeline creation is not supported'
+        )
+
     if pipeline.meta.original_pipeline_id is None:
         pipeline.check_name(session)
     else:
@@ -98,8 +113,14 @@ async def get_pipeline(
     name: str,
     version: Optional[int] = None,
     session: Session = Depends(service.get_session),
+    airflow: AirflowClient = Depends(airflow_utils.get_api_instance),
 ) -> Any:
     """Get latest pipeline from DB by name (and version)."""
+    if name.endswith(':airflow'):
+        dag_api = DAGApi(airflow)
+        name = ':'.join(name.split(':')[:-1])
+        dag = dag_api.get_dag(name, async_req=True).get()
+        return airflow_utils.dag2pipeline(dag)
     res = service.get_pipelines(session, name, version)
     if res:
         last_pipeline = sorted(res, key=lambda p: p.version)[-1]  # type:ignore
@@ -117,14 +138,27 @@ async def get_pipeline(
 async def get_pipelines(
     name: Optional[str] = None,
     session: Session = Depends(service.get_session),
+    airflow: AirflowClient = Depends(airflow_utils.get_api_instance),
 ) -> Any:
     """Gets all versions of pipeline from the DB.
     If 'name' is not specified, gets all pipelines."""
+    dag_api = DAGApi(airflow)
     if name is not None:
-        res = service.get_pipelines(session, name)
+        if name.endswith(':airflow'):
+            name = ':'.join(name.split(':')[:-1])
+            dag = dag_api.get_dag(name, async_req=True).get()
+            return [airflow_utils.dag2pipeline(dag)]
+        return [
+            pipeline.as_dict() for pipeline in service.get_pipelines(session, name)
+        ]
     else:
-        res = service.get_all_table_instances(session, dbm.Pipeline)
-    return [pipeline.as_dict() for pipeline in res]
+        dags = dag_api.get_dags(async_req=True).get()
+        return [
+            pipeline.as_dict() for pipeline
+            in service.get_all_table_instances(session, dbm.Pipeline)
+        ] + [
+            {"name": dag.dag_id, **dag.to_dict()} for dag in dags.dags
+        ]
 
 
 @app.post(
@@ -133,23 +167,58 @@ async def get_pipelines(
     response_model_exclude_unset=True,
     tags=["pipelines"],
 )
-def search_pipelines(
-    request: dbm.PipelineFilter,
+async def search_pipelines(
+    request: schemas.SearchFilters,
     current_tenant: Optional[str] = Header(None, alias="X-Current-Tenant"),
     session: Session = Depends(service.get_session),
+    airflow: AirflowClient = Depends(airflow_utils.get_api_instance),
 ) -> Any:
     """Returns a list of Jobs in line with filters specified"""
-    # todo: add airflow pipelines by rest api
-    query = session.query(dbm.Pipeline)
-    filter_args = map_request_to_filter(request.dict(), "Pipeline")
+
+    dag_api = DAGApi(airflow)
+
     try:
-        query, pag = form_query(filter_args, query)
+        query = session.query(dbm.Pipeline)
+        if request.name_pattern:
+            query = query.filter(dbm.Pipeline.name.like(f'%{request.name_pattern}%'))
+        pipelines_count = query.count()
+        query = query.offset(request.offset//2).limit(request.limit//2)
+        pipelines = query.all()
     except BadFilterFormat as e:
         raise HTTPException(
             status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
             detail=f"{e}",
         )
-    return paginate([element for element in query], pag)
+
+    try:
+        dag_filters = dict(
+            **request.dict(exclude_none=True, exclude={'name_pattern', 'limit', 'offset'}),
+            dag_id_pattern=request.name_pattern, limit=request.limit//2,
+            offset=request.offset//2
+        )
+        dag_filters = {k: v for k, v in dag_filters.items() if v is not None}
+        dags = dag_api.get_dags(**dag_filters, async_req=True).get()
+    except ApiException as e:
+        raise HTTPException(
+            status_code=status.HTTP_422_UNPROCESSABLE_ENTITY,
+            detail=f"{e}",
+        )
+
+    total_pages = ceil((dags.total_entries + pipelines_count) / (request.limit//2))
+    page_num = ceil((request.limit + request.offset) / request.limit)
+
+    pag = PaginationParams(
+        page_num=page_num,
+        page_size=request.limit or (len(dags.dags) + len(pipelines)),
+        min_pages_left=total_pages - page_num,
+        total=int(dags.total_entries + pipelines_count),
+        has_more=dags.total_entries + pipelines_count - request.offset // 2 * 2 - len(dags.dags) - len(pipelines) > 0,
+        page_offset=request.offset // request.limit,
+    )
+    dags = [airflow_utils.dag2pipeline(dag) for dag in dags.dags]
+    pipelines = [element for element in query]
+
+    return paginate(dags + pipelines, pag)
 
 
 @app.delete(
@@ -164,6 +233,13 @@ async def delete_pipelines(
     session: Session = Depends(service.get_session),
 ) -> Dict[str, str]:
     """Delete pipelines from db by name. All versions if not provided."""
+
+    if name.endswith(':airflow'):
+        raise HTTPException(
+            status_code=400,
+            detail='Airflow pipeline can not be deleted'
+        )
+
     res = service.get_pipelines(session, name, version)
     if res:
         service.delete_instances(session, res)
@@ -268,7 +344,7 @@ async def get_tasks_by_pipeline_id(
     tags=["execution"],
 )
 async def execute_pipeline_by_id(
-    pipeline_id: int,
+    pipeline_id: Union[int, str],
     args: List[schemas.InputArguments],
     job_id: Optional[int] = None,
     task_name: Optional[str] = None,
@@ -276,8 +352,26 @@ async def execute_pipeline_by_id(
     session: Session = Depends(service.get_session),
     x_current_tenant: Optional[str] = Header(None, alias="X-Current-Tenant"),
     token: TenantData = Depends(TOKEN),
+    airflow: AirflowClient = Depends(airflow_utils.get_api_instance),
 ) -> List[Dict[str, Any]]:
     """Schedule pipeline execution n times for n-given args."""
+
+    if isinstance(pipeline_id, str):
+        if not pipeline_id.endswith(':airflow'):
+            raise HTTPException(status_code=400, detail='Wrong pipeline_id')
+
+        dag_run_api = DAGRunApi(airflow)
+        dag_id = ':'.join(pipeline_id.split(':')[:-1])
+        dag_run_id = f'{dag_id}:{job_id}'
+
+        dag_run = DAGRun(
+            dag_run_id=dag_run_id,
+            conf={'files_data': [f.to_dict_for_airflow(job_id, x_current_tenant) for f in args]},
+        )
+        dag_run_api.post_dag_run(dag_id, dag_run, async_req=True).get()
+
+        return [{"id": dag_run_id}]
+
     res = service.get_table_instance_by_id(session, dbm.Pipeline, pipeline_id)
     if res is None:
         raise HTTPException(status_code=404, detail=NO_PIPELINE)
